switched loading argument s to be float

+++ Files/include/intro.h
@@ -4,7 +4,7 @@
 //put your function headers here
 //  Intro & Character Selection  /////////////////////////////////////
 void intro();
-void loading(int s);
+void loading(float s);
 void chooseCharacter();
 void printCharacterSheet(int currentChar);
 void archerSheet();
